pyproject.toml
[tool.poetry]
name = "gestorfinanciero"
version = "0.1.0"
description = "WIP"
authors = ["Ignacio <meloignacionicolas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
bcrypt = "^4.2.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
